chore: move group functions to helicity.decay
